Move viewmodels to separate package
